// frontend/src/api/model/githubCom4H1RZooraInternalDomainNotificationAudienceDTO.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Zoora API
 * REST API for the Zoora education platform.
 * OpenAPI spec version: 1.0
 */
import type { GithubCom4H1RZooraInternalDomainNotificationAudienceDTOType } from './githubCom4H1RZooraInternalDomainNotificationAudienceDTOType';

export interface GithubCom4H1RZooraInternalDomainNotificationAudienceDTO {
  class_id?: string;
  org_id?: string;
  role_id?: string;
  type: GithubCom4H1RZooraInternalDomainNotificationAudienceDTOType;
  /** @maxItems 500 */
  user_ids?: string[];
}